+++ src/Http/Middleware/NotVerified.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Posio\CabinetKit\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+
+class NotVerified
+{
+    /**
+     * Handle an incoming request.
+     *
+     * @param  \Closure(\Illuminate\Http\Request): (\Symfony\Component\HttpFoundation\Response)  $next
+     */
+    public function handle(Request $request, Closure $next)
+    {
+
+        if ($request->user() && !$request->user()->hasVerifiedEmail())
+            return redirect(route("verification.notice", absolute: false));
+
+        return $next($request);
+    }
+
+
+}